Move datetime input widget to separate app
