+++ meinberlin/apps/datetimefield/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class Config(AppConfig):
+    name = 'meinberlin.apps.datetimefield'
+    label = 'meinberlin_datetimefield'
